Add Multiple HHS Subdomains and 1 SBA Subdomain
This pull request aligns with the pull request of the equivalent GSA/Data/other-websites file, which I do not believe GSA is watching or maintaining.  We should probably start managing our own list as started here and point the gatherer at this list instead of https://github.com/GSA/data/blob/master/dotgov-websites/other-websites.csv.  This would be in alignment with what we did in https://github.com/cisagov/gatherer/pull/51 (the line in https://github.com/cisagov/gatherer/blob/develop/gather-domains.sh in this case which would need to be modified would be line 99, I think).

----

HHS requested we add a large handful of subdomains not showing up in their HTTPS/Trustymail reports. The original list has been compared to the list we are currently scanning, and those that weren't already being scanned have been added here.

HHS sent us a long list of domains to add to their BOD 18-01 reports. We reviewed all the non-.gov base domains for association that it was either managed by or specifically on HHS' behalf, resulting in the smaller number of adds you see at cisagov/scan-target-data#48.  These are the subdomains for those that were included in the list.

SBA has requested that we add disasterloanassistance.sba.gov so that it shows up in their Trustymail and HTTPS reports.  I verified it is not currently being picked up in the sources gathered by double checking for its presence in their report.